+++ Project/DataModels/MenuItemModel.cs
@@ -3,13 +3,13 @@ public class MenuItemModel : IModel
     public Int64 Id { get; set; }
     public string Name { get; set; }
     public string Category { get; set; }
-    public decimal Price { get; set; }
+    public double Price { get; set; }
 
     public MenuItemModel(
         Int64 id,
         string name,
         string category,
-        decimal price
+        double price
     )
     {
         Id = id;
